Fin3

+++ Practica3/Complejos_Java_TablaDeSimbolos/Complejo.java
@@ -1,13 +1,13 @@
 public class Complejo {
-  private double real;
-  private double img;
+  private float real;
+  private float img;
 
-  public Complejo(double real, double img) {
+  public Complejo(float real, float img) {
     this.real = real;
     this.img  = img;
   }
 
-  public double getReal() {return real; }
+  public float getReal() {return real; }
   
-  public double getImg() {return img; }
+  public float getImg() {return img; }
 }
